+++ extensions/dash/src/components/DashResult.tsx
@@ -1,4 +1,5 @@
 import { ActionPanel, List, OpenAction } from "@raycast/api";
+import { decode } from "html-entities";
 import { DashResult } from "../util/useDocsetSearch";
 
 export default function ({ result, index }: { result: DashResult; index: number }) {
@@ -6,8 +7,8 @@ export default function ({ result, index }: { result: DashResult; index: number 
   return (
     <List.Item
       key={result["@_uid"]}
-      title={result.title.toString()}
-      subtitle={subtitle}
+      title={decode(result.title.toString())}
+      subtitle={decode(subtitle)}
       accessoryTitle={docsetName}
       icon={result.icon}
       actions={