+++ extensions/gitlab/src/components/project_search.tsx
@@ -141,11 +141,13 @@ export function useSearch(query: string | undefined): {
   const [error, setError] = useState<string>();
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
-  let cancel = false;
-
   useEffect(() => {
+    // FIXME In the future version, we don't need didUnmount checking
+    // https://github.com/facebook/react/pull/22114
+    let didUnmount = false;
+
     async function fetchData() {
-      if (query === null || cancel) {
+      if (query === null || didUnmount) {
         return;
       }
 
@@ -155,15 +157,15 @@ export function useSearch(query: string | undefined): {
       try {
         const glProjects = await gitlab.getProjects({ searchText: query || "", searchIn: "title" });
 
-        if (!cancel) {
+        if (!didUnmount) {
           setProjects(glProjects);
         }
       } catch (e: any) {
-        if (!cancel) {
+        if (!didUnmount) {
           setError(e.message);
         }
       } finally {
-        if (!cancel) {
+        if (!didUnmount) {
           setIsLoading(false);
         }
       }
@@ -172,7 +174,7 @@ export function useSearch(query: string | undefined): {
     fetchData();
 
     return () => {
-      cancel = true;
+      didUnmount = true;
     };
   }, [query]);