+++ extensions/kill-process/src/index.tsx
@@ -7,8 +7,9 @@ export default function ProcessList() {
   const [query, setQuery] = useState<string | undefined>(undefined);
   const shouldIncludePaths = (preferences.shouldSearchInPaths?.value as boolean) ?? false;
   const shouldPrioritizeAppsWhenFiltering = (preferences.shouldPrioritizeAppsWhenFiltering?.value as boolean) ?? false;
+  const shouldShowPID = (preferences.shouldShowPID?.value as boolean) ?? false;
 
-  useEffect(() => {
+  const fetchProcesses = () => {
     exec(`ps -eo pid,pcpu,comm | sort -nrk 2,3`, (err, stdout) => {
       if (err != null) {
         return;
@@ -34,6 +35,10 @@ export default function ProcessList() {
 
       setState(processes);
     });
+  };
+
+  useEffect(() => {
+    fetchProcesses();
   }, []);
 
   const fileIcon = (process: Process) => {
@@ -78,8 +83,8 @@ export default function ProcessList() {
           return nameMatches || (shouldIncludePaths && pathMatches);
         })
         .sort((a, b) => {
-          // If this flag is true, we bring apps to the top.
-          if (shouldPrioritizeAppsWhenFiltering) {
+          // If this flag is true, we bring apps to the top, but only if we have a query.
+          if (query != null && shouldPrioritizeAppsWhenFiltering) {
             if (a.type === "app" && b.type !== "app") {
               return -1;
             } else if (a.type !== "app" && b.type === "app") {
@@ -87,7 +92,7 @@ export default function ProcessList() {
             }
           }
 
-          // Otherwise, we leave the order as it is.
+          // Otherwise, we leave the order as is.
           return 0;
         })
         .map((process, index) => {
@@ -96,13 +101,19 @@ export default function ProcessList() {
             <List.Item
               key={index}
               title={process.name}
-              subtitle={process.id}
+              subtitle={shouldShowPID ? process.id : undefined}
               icon={icon}
-              accessoryTitle={process.cpu}
+              accessoryTitle={`${process.cpu}%`}
               actions={
                 <ActionPanel>
                   <ActionPanel.Item title="Kill" icon={Icon.XmarkCircle} onAction={() => killProcess(process)} />
                   {copyToClipboardAction(process)}
+                  <ActionPanel.Item
+                    title="Reload"
+                    icon={Icon.ArrowClockwise}
+                    shortcut={{ key: "r", modifiers: ["cmd"] }}
+                    onAction={() => fetchProcesses()}
+                  />
                 </ActionPanel>
               }
             />