+++ extensions/open-with-app/src/index.tsx
@@ -5,6 +5,7 @@ import {
   Application,
   getApplications,
   closeMainWindow,
+  popToRoot,
   showHUD,
 } from "@raycast/api";
 import { useState, useEffect } from "react";
@@ -55,7 +56,7 @@ async function getFinderSelection():Promise<string[]> {
 
   return (response === '')
     ? []
-    : response.split('\r').map((item) => item.replace(/"/g, '\\"'));
+    : response.split('\r');
 }
 
 function ApplicationsListItem(props: { application: Application }) {
@@ -71,15 +72,16 @@ function ApplicationsListItem(props: { application: Application }) {
           <ActionPanelItem
             title={`Open with ${application.name}`}
             onAction={async () => {
-              closeMainWindow();
               const selectedItems = await getFinderSelection();
               if (selectedItems.length === 0) {
                 await showHUD(`⚠️  No Finder selection to open.`);
               } else {
                 selectedItems.forEach((item) => {
-                  execSync(`open -b ${application.bundleId} "${item}"`);
+                  execSync(`open -b ${application.bundleId} "${item.replace(/"/g, '\\"')}"`);
                 })
               }
+              closeMainWindow();
+              popToRoot({ clearSearchBar: true });
             }}
           />
         </ActionPanel>