+++ extensions/visual-studio-code-project-manager/src/types.ts
@@ -1,0 +1,18 @@
+export interface ProjectEntry {
+  name: string;
+  rootPath: string;
+  tags: string[];
+  enabled: boolean;
+}
+
+export interface GitCachedProjectEntry {
+  name: string;
+  fullPath: string;
+}
+
+export interface Preferences {
+  projectManagerDataPath: string;
+  groupProjectsByTag: boolean;
+  terminalAppPath: string;
+  gitClientAppPath: string;
+}