+++ docs/changelog.md
@@ -1,8 +1,38 @@
 # Changelog
 
+## 1.25.4 - 2021-11-11
+
+### 💎 Improvements
+
+* Updating of items and submenus while the action panel is open
+* Supporting all convenience actions with primary shortcut (cmd + enter) on form views
+* Better error handling when the API cannot be loaded after failed app updates
+
+### 🐞 Bugfixes
+
+* Loading indicator in detail views when used in a navigation stack
+
+## 1.25.2 - 2021-10-28
+
+### 💎 Improvements
+
+* Improved ActionPanel updating performance
+
+### 🐞 Bugfixes
+
+* `searchBarPlaceholder` updates when using the list in a navigation stack
+* Wrong action panel actions when popping back in a navigation stack
+* Empty state flickering when updating the `isLoading` property in lists
+* Accessory and subtitle label truncation in lists
+* Icon from assets tinting on dynamic theme changes
+* Dynamic removal of form elements
+* Open actions leading to Node env vars being set for the opened application
+* Some extensions not getting loaded for a particular Node setup
+* Local storage values being lost when extensions are automatically updated
+
 ## 1.25.1 - 2021-10-20
 
-## 🐞 Bugfixes
+### 🐞 Bugfixes
 
 * Fixed configuring `tintColor` for icons in `ActionPanel` and `Form.Dropdown`
 * Fixed displaying submenu icons from local assets